backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
grpcio
grpcio-tools
protobuf